requirements.txt
asgiref==3.8.1
certifi==2024.7.4
channels==4.1.0
charset-normalizer==3.3.2
Django==5.0.7
djangochannelsrestframework==1.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
idna==3.7
packaging==24.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
whitenoise==6.7.0
